+++ src/components/Main/Selector/SelectorForm/Select/SelectValue.jsx
@@ -1,10 +1,11 @@
 import { useContext, useEffect, useState } from "react";
 import { SelectorContext } from "../../../../../store/selector-context.jsx";
-import { $BuilderValues } from "./BuilderValues.styles.jsx";
+import $Select from "./Select.styles.jsx";
 import { apiCallToGetFilterValues } from "../../../../../api/Api.jsx";
 
-export default function BuilderValues() {
+export default function SelectValue() {
   const [filterValues, setFilterValues] = useState([]);
+  const [firstOption, setFirstOption] = useState("");
   const { tempWhereCriteria, setTempWhereCriteria } =
     useContext(SelectorContext);
 
@@ -14,7 +15,11 @@ export default function BuilderValues() {
         await apiCallToGetFilterValues(tempWhereCriteria.key).then(
           (resData) => {
             setFilterValues((_prevFilterValues) => resData);
-            setTempWhereCriteria({ ...tempWhereCriteria, value: resData ? resData[0].name : null});
+            setTempWhereCriteria({
+              ...tempWhereCriteria,
+              value: resData ? resData[0].name : null,
+            });
+            setFirstOption(resData ? resData[0].name : "");
           }
         );
       } catch (error) {
@@ -22,20 +27,18 @@ export default function BuilderValues() {
       }
     }
     fetchData();
-    
   }, [tempWhereCriteria.key]);
 
   function handleWhereCriteriaChange(event) {
+    setFirstOption(event.target.value);
     setTempWhereCriteria({ ...tempWhereCriteria, value: event.target.value });
   }
 
   return (
-    <$BuilderValues>
-      <select onChange={handleWhereCriteriaChange} defaultValue={"0"}>
-        {filterValues.map((filterValue) => (
-          <option value={filterValue.id} key={filterValue.id}>{filterValue.name}</option>
-        ))}
-      </select>
-    </$BuilderValues>
+    <$Select onChange={handleWhereCriteriaChange} value={firstOption} disabled={filterValues.length === 0}>
+      {filterValues.map((filterValue) => (
+        <option key={filterValue.id}>{filterValue.name}</option>
+      ))}
+    </$Select>
   );
 }