+++ src/components/Main/Selector/SelectorForm/SelectorForm.styles.jsx
@@ -1,11 +1,11 @@
 import { styled } from "styled-components";
 
-export const $SelectorBuilder = styled.div`
+export const $SelectorForm = styled.form`
   margin: 25px 100px 25px 0;
   display: flex;
   justify-content: flex-start;
   column-gap: 10px;
   align-items: center;
-`;
+`
 
-export default $SelectorBuilder;
+export default $SelectorForm;